requirements.txt
absl-py==2.1.0
astunparse==1.6.3
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1734473278428/work/certifi
charset-normalizer==3.4.1
click==8.1.8
filelock @ file:///croot/filelock_1700591183607/work
flatbuffers==25.1.24
gast==0.6.0
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1737760107953/work
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml-dtypes==0.4.1
mpmath @ file:///croot/mpmath_1690848262763/work
namex==0.0.8
networkx @ file:///croot/networkx_1737039604450/work
nltk==3.9.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=5f96f274d410a1682519282ae769c877d32fdbf171aa8badec7bf5e1d3a1748a
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow @ file:///croot/pillow_1738010226202/work
protobuf==5.29.3
Pygments==2.19.1
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sympy @ file:///croot/sympy_1738108488918/work
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.1
urllib3 @ file:///croot/urllib3_1737133630106/work
Werkzeug==3.1.3
wrapt==1.17.2
